Cleanup and renamed CopybookDownloadService to ExternalAPIsService
